// src/api/swagger/models/TransactionRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TransactionRequest = {
    userId?: string;
    partnerId?: string;
    transactionCategoryId?: string;
    amount?: number;
    transactionType?: string;
    status?: string;
    note?: string;
    attachment_id?: Array<string>;
};